feat: stock the pond
was looking for a better way in the example app to define what actions are applied to a fish. Since the entire app is generated based on the dir, the challenge is how to add defined behavior to things determined at runtime. This feels better than before.

Now you can define fish to feed into a pond. See comments.