feat: auto define templates
Now, instead of parsing template via a collection of file paths I will read them, wrap in expected define syntax, and the template will parse a string of the buffer.

I like this change because it enables me to make templates without worrying about defining their name.  For example if I wanted to name blogs by their day it was annoying to have to define same as file name. Plus what if I want to change those file names later. Not so painful now.

The place this could be issue is the sardine template usages but I cannot automate everything. This does not _break_ than, just enforces a convention.