feat: ponds flow inherit asset fish
I had an idea of wanting to connect ponds to each other. This is the first iteration of that. Now, I can make a pond just for assets, and have it flow into the main pond.  A catch is that only anchovy and clown fish are what make sense for this feature.

Since I make decisions based on dir structure it needs to be clear. No murky water. Now we can break assets out however we want and have them apply to our "app" as a whole.