abstraction: html element
so yesterday at lunch I read a bit more of that golang mistakes book, and liked a quote. It was something like doing more with less and how powerful simple abstraction can be. I love this concept and felt the previous way I was approaching a powerful ux framework was too narrow minded. That is okay though, because it gave me the chance to do this. Look at it! I mean, sure its kind of a virtual dom. BUT that is cool to me, and I cna see all kinds of ways to expand on this way more than the previous method I was using.

So the actual changes? Oh yeah. I now have a single "element" template that is recursive. A simple struct with a map of attributes. So freaking cool imo.

Also, this lets me encapsulate the translation form go to HTML away from the consumer, e.g. not needing the template func map anymore (tackle)