abstraction: attribute from element
So I was thinking about it, and not sure if I want to keep methods on the attribute part. Its just a map[string]string after all - not too complex. However, I did think that if I was to keep the methods related to an attribute, then that is where they belong - not on an element. Since the concept of an element should not be limited to the idea of an input where value is relevant, but an table element value is not. Also moved some time print funcs into their own home. Not sure what belongs there now, but seemed appropriate. Finally, removed the variations of a text input in favor of just one that takes in the kind since all the funcs where the same anyway.