rename: element to bridge
I have always had this gripe with web development that the server and the app have to duplicate work in different ways. Starting in Angular I used to never do server validation. I loved angular reactive forms and how well defined they where - at least compared to other frameworks I had explored. I thought that was all I needed. Wrong. So, naturally, my backend was in JavaScript. Once I decided it needed validation, the extent of that was checking for undefined and trusting the request after that. Dozens of middleware and hundreds of lines that still did not accomplish what I wanted. Everything felt bad.

I have tried making some other tools like a tool that writes the front and back end for me given a configuration. So long as the config is the same if should work. However, I found that solution to be unrealistic as generating boilerplate code never solved the problems, and often didn't even get me as far as I had hoped.

All this to say, I think that bridge is the most appropriate name for what this package is designed to accomplish. I noted some more about the package name in its comment.