feat: parse r form based on html el
So now I have been working on what should a bridge between html a go for a form look like. So far this is what I got. Some methods to help traverse an element tree to aid in manipulating (via pre-setting values) or understand (via gathering keys) it. The primary start of the show I wrote a test for. It takes in a request (generated by the html element (aka form)) and sees what form the user was filling out to determine which values are relevant to parse from the request.

I think its pretty cool and its a step towards even cooler things. Also, various other minor fixes and changes. One notable is the html element template is now a no-whitespace one liner. This was to remedy issues with the text area value being based on how its rendered. That said, its also a minor win as there is no more wasted bytes.