add: local/global generics & feat: anchor el
So this commit was not meant to be a generics addition. But that is what it turned into after I started working on my idea for an anchor. An anchor that can be defined before the application is ran to grantee links work as will get to a page. So solving as many problems that a site can run into, before it is ran.

Main thing of note is that a pond has a struct for its "global" data, and a fish has a struct for its "local" data. Global in in my case was to ensure all pages have access to the anchor tags, to be used by the nav template.

Several other changes cam into play as I worked on getting the demo app functional too.
- sort routes before registration to ensure consistency
- sort document head links (css/js) so the page is always rendered the same.
- fixes to the gobble function
- all methods on a fish or pond are now functions as methods on a struct do not support generics - very sad about this
- sardines handler will also use global fish to render template
- took parent (pond) off of fish to prevent fish needing generic type of pond too. Plus its parent was only used when casting a line and reeling, so I just pass it in instead
- changes how the hashing works to be a version query param. I saw that somewhere and liked it since it does not obfuscate the file name when developing.
  - using this also allowed me to rename the htmlx file since its "v" is its has. No need for version in file name.