feat: new table el
a function to provide an html element, a table. Takes in csv data for now since all it does is print the table. I want to build on this element to have the expected features of a table, like pagination, sort, filter... but this is the first step towards that