feat: stateful element
So I started working for several hours on a cool stateful table. And was ending up with something that kind of worked, but felt bad. I was designing a function so specific to what I needed it was not useful for anything else. So I stashed that for later in favor of this change.

It is also a lot of learning for me to see how state is managed in a ssr application. What I landed on was simple but sweet: hidden html inputs. It is small, but i think has great potential for making the original goal of elements like a stateful table without backing myself into a corner.