feat: delete element
Similar to the find features I found myself needing a way to delete elements easily. Main example is wanting to use the NewInputSelect func to create a dropdown, but not wanting the label that comes with it. Instead of making a "include label" param on that fn I will just enable deleting from an element as that seems more powerful and versatile.

test included and passed. Not gonna like this hurt my brain a little, thank goodness for the tests and being able to step through code.