fix: utilize make fn capacity
so this is not really a fix, but kind of. At work I was discussing with a co worker and he mentioned how capacity and length are different. My Javascript brain was confused, but then it started to make sense. Knowing this, I wanted to add it to my code too. Everything seems to work fine.

Still torn on giving things 0 capacity by default. I know its a trade off when giving capacity that may not be used. For now I kept some things at 0 unless I knew how big it should be.