restructure: bridge from example to pkg & mackerel fish
I have been spending a lot of time on the bridge package, and decided that it belongs part of my framework, not just an example app. Lots of moving things around, a few change:

- fix: element package set value and checked/selected index
- add: mackerel fish. Similar to sardine but not served and is predefined bytes (only mackerel for now is element one). I don't plan on having many of these. A 'system' fish that uses pre-defined bytes
- rm: element template
- add: when creating a pond it adds my element mackerel automatically
- mv: internal -> pkg since that is more appropriate for a library
- rm: water example. It was meant to show how state can be preserved using a form. This is now better demonstrated by the stateful table.
- rename package: internal -> pkg/aquatic
- rm: anchor. Was not happy with this as it felt gross / over complicated for an anchor. I like the idea and may try it again later in another way.

Overall not anything too big. Mostly changes to get bridge into pkg. Glad to have the bridge part of the library now. Was not sure how to approach that for a while since it was a pond at first being derived from a folder. Pretty happy with what I landed.