improve: pre alloc and cache fish template bytes
Now when reading the template of a fish I will pre-alloc the size needed to wrap its contents into the define syntax. Also that read never needs to change for duration of the app running, so I just save it with the fish so later when we need its template we can just access it instead of re building it.

Also this was my first time using the slice start end colin syntax, I think it is cool. Learned that to copy I need to copy into len, not cap.