improve: tuna handle func + bobber
With the same goal as other commits tonight I an trying to improve how I manage memory with fish. Now there is a bobber for a fish, it just holds head links for a tuna. Between coral, reef, and bobber the tuna handler feels a lot better. Stuck with a buffer and write to it though since exec template size is not pre defined.