add: test for element
working on improving ability to make it 'just work' within reason. For this commit that means being able to set the value of a new html input without worrying about finding the input element within the div. Because my new input functions return a div with a label and input. Also handling text area setting value.

Decided to keep the setting value of for 'select' element separate. Because the occurrence is for targeting an element (say the 3rd 'select' element) which works fine for say a checkbox as it is the nth input is the element we wanted to modify. But that did not work for select as we want the option of that select to be modified.

Tweak to working on my 'like' fns to keep consistent.