add: drag-drop example
One of the limitations is live user interaction. Usage of htmlx helps a ton - but is not a all in one solution. There is no escaping the javascript it seems.

Something I look for when deciding what tools or libraries to use is the features it has, and how does it stand out from the others to save me time solving a business problem.

One of the first things I used to look for was a powerful table with all the shiny features. If a framework or library had that, it was immediately more viable for getting me to a finished product. So that was one of the first things I wanted to prove to myself. "Can the tool I am making do what I expected of others?". The answer was yes as far as tables, see commit cd5e8d278c71a3be365d5d9b350368ea032213a0

Moving on to this commit. Another use case that I find in more involved web apps is drag-drop features, typically dashboards. Angular cdk comes to mind.  Like the table, if I can use the tool I made to do these things, then that is a measure of success in my mind.

Anyway, this commit proves my flow of Define>Populate>Modify works with a drag-drop use case, albeit with the use of a little javascript.

Not 100% sure on the solution, I see pros and cons... the biggest con being lag between client and server. However that same lag is present in spa, just more convoluted.

Oh and a fix so school fish have clown hash in head.