feat: reflection + validation
This turned into on of the biggest commits - unexpectedly too. Lots of learning here too which I am proud of.

I knew that I wanted to start making validation since I already had a `ParsedForm` struct that could look at en element and its children to see what the value was. So, I got started on leveling up that struct.

I spent a few hours writing and rewriting ways to constantly check the `map[string]string` that is an elements attributes, parse them out, and compare. Value, min, max, required, minlen, maxlen, pattern... Whatever I wrote felt wrong. Lots of repeating myself, maybe not the code but the patterns I was following. So, I decided to look into reflections.

Reflections was something I had heard about from a co-worker. He had implemented it and I knew it was cool, but the feature felt out of my league. But today I gave it a shot. After reading the laws of reflection and dinking around in a scrap project I had landed on my goal for how I would implement with this project.

After starting it kind of fell into place, the `map[string]string` that had given me so much trouble, suddenly was a non issue. The validation struct and being able to parse attribute strings into it was so much better than what I had previously been doing. I was finally able to make a validation function I was mostly happy with. After writing an annoying amount of tests for validation (albeit small ones), I felt confident in my decision to go with reflections.

Perhaps I should have committed there, but the can of worms was open and there was so much I wanted to do with my new super powers. After some tweaks to the validation solution (isolating the ability to set a string to a reflect value) I was able to reuse it to revisit previous solutions that felt gross. Take a look ath the `element.go` and `form.go` files. See how many repeated ideas where consolidated! I love that.

Here is a list of what all happened:
- removed AttributeKey type in favor of string. Since I was casting all the time anyways and it took away from the point of HTMLElement being simple
- removed form parse functions in favor of `ValueOf` which uses reflections
- removed element parse functions in favor of `ElementValue` which uses reflections
- modified example app to use new functions
- renamed functions to follow a convention. Because the use of generic I cannot have func bound to struct anymore. Making me feel like it was getting lost in all the functions within this package. So for now, if a function would have been on a struct but cannot due to a generic, I ensured the first word of the fn was consistent to that struct. So `ElementFoo` and `ElementBar` not `FooElement`. I was thinking about breaking bridge out into smaller packages but will have to think on that some more.

Just a side bar about reflections too. The concept of [value, concrete type] reminds me of why I love go, it is simple concepts made powerful. While reflections are still kinda scary, it felt great being able to learn about it and mostly understand it - at least enough to start using it.