feat: manipulate el class and style
After tinkering with more technical things recently I wanted to work with something else (a little less technical) that I enjoy: making website look pretty. This is because the tradeoff when using HTMLElement struct is that styling and visualizing the ux in my head is much more difficult than with the templating of classic html and css.

So I desired tooling around manipulating the style and class attributes. It was a fun simple enough solution that I had been putting off until now.

Coolest thing I learned was the SplitN function. I have ran into that exact need in other projects, and thing it is super amazing that it is something that exists in the strings package. Good job go, thank you.