fix: selection inputs use stringer
So I did not realize it at the time, but I had just reinvented the something that already existed. Having read a bit more about it in the a book it mentioned how implicit satisfaction of interfaces. Such as stringer - which was why I needed printable. So we can use that instead.