add: HTMLInput, use less reflection
So after some more thinking and weighing the pros and cons I decided to not use reflection for extracting a value form an html input. The main reason was the strconv package. I thought to myself: if it is good enough for that package to make functions, it is good enough for me.

The reflection based functions ElementValue and ValueOf  where unessesary over compilation of extracting a value from an html element. However I think there is still value in the attributes to struct reflection, just not here.

In a similar note, I wanted to keep more organized. So, HTMLInput is born. When doing this I want to make sure that it remains an HTMLElement, just with extra features. This ensures that I can both keep the abstraction and simple nature of an element, while also including relevantly scoped features. I am like this approach and want to expand on it more.