feat: table el & mods
Several hours, and a few days break, went into this. At first I started with wanting a table to be similar to an HTMLInput, in the since that it is just an HTMLElement, but with special fns. However, I ran into issue (will get into later). It was a simple problem to understand but I had backed myself into a corner trying to keep ultra simple abstraction. So, I took a few days break.

Having come back to it with fresh eyes and having learned more about the role of state vs config I gave it another swing. Keep in mind, the main flow is 3 step. 1: Defined the element. 2: Populate it via request to preserve state and accept changes to state. 3: Modify the element. It was after step 2 that I was previous hung up on when trying to make the Constraint fn (help with queries). The form state was restored form the request, but I had no way of knowing the value of a select I had just made! Because all I care about are Stringer satisfaction to print the html, and all values where indexes. So instead of the row value I was getting row count index...

Allowing myself to broaden what a HTMLTable was enabled me to make something I am much prouder of. For example :
- By storing the config of how a table was made I was able to make more informed decisions about manipulating it. This is helpful for repeat static info such as the hx post target for swapping.
- By storing state at the I was able to
  - not have to re-lookup elements in the table. Since in the define step I make the elements, I can just store references ot them to look at in the modify step. Much better.
  - export specific feature for consumer, e.g.: RecordCount so the pagination has good data to go off of. Which can be set at prior to step 3 (modify) which is logical since we make our query after step 2 using the Constraint fn.

Specifics are commented if curious.