rm: bridge and simplify bait
so its been a while. But I started playing with side projects again and wanted to use my gofish packages to get it started. And realized that I was not a fan of using what I had made. More specifically I did not having to build a ui without html and css - which I enjoy (in small doses lol). The bridge package conflicted with this.

While the fish and ponds focused on streamlining html templating for a solid standalone frontend, the bridge focused on a frontend-backend pairing that almost had to live together. And used a generic "element" that to render anything was overlapping with a sardine in a way that felt wrong.

It broke my ability to just make a frontend, and was forcing me into ux decisions that my own library deemed correct. This became more of an issue once the bridge started handling more complex such as the table. For example, I added filtering and sorting to headers like I have done to most all other crud tables in other projects. But recently I have liked the idea of a tables filtering and sorting to be on the left of screen - similar to online retails. But since I made that decision in my table bridge element, it was a headache to make a relatively simple change.

That said, I think that the bridge package was fun and exercise and cool concept that I may revisit later. But as far as where I think this project should be I want to refine the way fish and ponds work.

Also, I did not like the use of generics for data being passed into templates. For me, it did not make using data in a template any more friendly, and added complexity to a pond requiring a generic structure with optional fields, and each template would only care about one of those fields. So for now I will use any and may re think it later.

And since we are not using generics anymore I can re-attach the functions as methods to structs - since I like methods.